requirements.txt
python==3.9.*
opencv-python==4.5.5
numpy==1.21.6
scikit-learn==1.0.2
tensorflow==2.8.0
pillow==9.0.1
psycopg2-binary==2.9.3
protobuf==3.19.6